+++ lib/ui/window/viewport_metrics.h
@@ -14,7 +14,7 @@ struct ViewportMetrics {
   ViewportMetrics(double p_device_pixel_ratio,
                   double p_physical_width,
                   double p_physical_height,
-                  int32_t p_physical_touch_slop);
+                  double p_physical_touch_slop);
   ViewportMetrics(double p_device_pixel_ratio,
                   double p_physical_width,
                   double p_physical_height,
@@ -30,7 +30,7 @@ struct ViewportMetrics {
                   double p_physical_system_gesture_inset_right,
                   double p_physical_system_gesture_inset_bottom,
                   double p_physical_system_gesture_inset_left,
-                  int32_t p_physical_touch_slop);
+                  double p_physical_touch_slop);
 
   double device_pixel_ratio = 1.0;
   double physical_width = 0;
@@ -47,7 +47,7 @@ struct ViewportMetrics {
   double physical_system_gesture_inset_right = 0;
   double physical_system_gesture_inset_bottom = 0;
   double physical_system_gesture_inset_left = 0;
-  int32_t physical_touch_slop = -1;
+  double physical_touch_slop = -1.0;
 };
 
 bool operator==(const ViewportMetrics& a, const ViewportMetrics& b);
